requirements.txt
attrs==19.1.0
coverage==4.5.4
dj-database-url==0.5.0
Django==2.2.4
django-extensions==2.2.1
django-rest-auth==0.9.5
djangorestframework==3.10.2
djangorestframework-simplejwt==4.3.0
filelock==3.0.12
graphviz==0.11.1
importlib-metadata==0.19
packaging==19.1
pluggy==0.12.0
psycopg2==2.8.3
py==1.8.0
pygraphviz==1.5
PyJWT==1.7.1
pyparsing==2.4.2
python-decouple==3.1
pytz==2019.2
six==1.12.0
sqlparse==0.3.0
toml==0.10.0
tox==3.13.2
virtualenv==16.7.3
zipp==0.5.2
